+++ src/ZplCommand/MapClear.php
@@ -1,0 +1,23 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Janisvepris\ZplBuilder\ZplCommand;
+
+use Janisvepris\ZplBuilder\Util\BoolToStr;
+use Janisvepris\ZplBuilder\ZplCommand;
+
+readonly class MapClear implements ZplCommand
+{
+    public const string COMMAND = '^MC';
+    public const string FORMAT = '%s';
+
+    public function __construct(
+        private bool $clear,
+    ) {}
+
+    public function __toString()
+    {
+        return self::COMMAND . sprintf(self::FORMAT, BoolToStr::conv($this->clear));
+    }
+}